Rework custom cache notification handlers, move index rebuilding notification handler to the in-memory provider, add a few more features to the search playground

+++ src/Umbraco.Cms.Search.Core/Services/ContentIndexing/DraftContentChangeStrategy.cs
@@ -86,6 +86,11 @@ public class DraftContentChangeStrategy : IDraftContentChangeStrategy
 
     private async Task RemoveFromIndexAsync(IndexInfo[] indexInfos, IReadOnlyCollection<Guid> keys)
     {
+        if (keys.Count is 0)
+        {
+            return;
+        }
+
         foreach (var indexInfo in indexInfos)
         {
             await indexInfo.IndexService.DeleteAsync(indexInfo.IndexAlias, keys);
